Moving SafeLibraryHandle to the Interop.Windows.Handles namespace, migrating usages
